scripts/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==0.4.3
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1694046349000/work
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
blobfile==2.0.2
braceexpand==0.1.7
cachetools==5.3.1
carvekit-colab==4.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.1.1
cycler==0.11.0
datasets==2.4.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1695534305529/work
decorator==4.4.2
defusedxml==0.7.1
dictionaries==0.0.2
diffusers==0.12.1
dill==0.3.5.1
docker-pycreds==0.4.0
einops==0.3.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastapi==0.103.1
fastcore==1.5.29
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.12.4
fire==0.4.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
gradio==3.21.0
grpcio==1.58.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.2
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.1.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1693880262622/work
ipython==8.15.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1690896916983/work
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1693317508789/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1695827927545/work
jupyterlab-pygments==0.2.2
kaleido==0.2.1
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.3
linkify-it-py==2.0.2
llvmlite==0.41.1
loguru==0.7.2
lovely-numpy==0.2.9
lovely-tensors==0.1.15
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
multidict==6.0.4
multiprocess==0.70.13
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
nibabel==5.1.0
numba==0.58.1
numpy==1.23.3
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
orjson==3.9.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.1.1
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathtools==0.1.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
plotly==5.13.1
proglog==0.1.10
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.24.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1695367186825/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pudb==2019.2
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-flags==1.1.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
pydecor==2.0.1
pyDeprecate==0.3.1
pydicom==2.4.3
pydub==0.25.1
pyGenericPath==0.2.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
Pympler==1.0.1
pyopencl==2023.1.4
pyparsing==3.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-multipart==0.0.6
pytools==2023.1.1
pytorch-lightning==1.4.2
pytranskit==0.2.9
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1695384318214/work
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.3.1
scipy==1.9.1
seaborn==0.13.0
sentry-sdk==1.31.0
setproctitle==1.3.2
shapely==2.0.1
sigpy==0.1.0
SimpleITK==2.3.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SSIM-PIL==1.0.14
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
streamlit==1.26.0
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.9.18
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1+cu113
torch-fidelity==0.3.0
torchgeometry==0.1.2
torchmetrics==0.6.0
torchvision==0.13.1+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1695373443748/work
tqdm==4.66.1
traitlets==5.10.0
transformers==4.22.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==1.26.16
urwid==2.2.0
uvicorn==0.23.2
validators==0.22.0
Vapory==0.1.2
wandb==0.15.10
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.7
xxhash==3.3.0
yarl==1.9.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
